% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucdpged}
\alias{ucdpged}
\title{UCDP Georeferenced Event Dataset (GED) sample}
\format{
a tibble with 1000 rows and 49 columns
}
\source{
<https://ucdp.uu.se/downloads/>
}
\usage{
ucdpged
}
\description{
A sample of the UCDP Georeferenced Event Dataset (GED) from the 2023 data release. The data contains information about the date, location, and type of conflict events. The data is a sample of the full dataset, which can be downloaded from the UCDP website <https://ucdp.uu.se/downloads/>.
}
\keyword{datasets}
